pyproject.toml
[tool.poetry]
name = "refactor-stats-maker"
version = "0.4.1"
description = ""
authors = ["Tiago Pereira <tiago.pereira@infraspeak.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
codeowners = "^0.6.0"
ripgrepy = "^2.0.0"
plotext = "^5.2.8"
pyperclip = "^1.8.2"
gitpython = "^3.1.40"
platformdirs = "^3.11.0"
halo = "^0.0.31"
black = "^23.10.1"
pytest-watch = "^4.2.0"
jira = "^3.5.2"
configparser = "^6.0.0"
more-itertools = "^10.1.0"
click = "^8.1.7"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
"refactor-stats-maker" = "refactor_stats_maker.__main__:run"
